Perform minor clean up and improvements
