+++ NorrisFacts/Service/Provider/ChuckNorrisAPI/CategoryProvider.swift
@@ -11,7 +11,6 @@ import Moya
 import RxMoya
 import RxSwift
 
-// TODO: Use enum for retryOnError
 protocol CategoryProviderProtocol {
     func fetchCategories(scheduler: SchedulerType, retryOnError: Bool) -> Single<[Category]>
 }
